requirements.txt
fastapi~=0.89.1
uvicorn~=0.20.0
pydantic~=1.10.4
python-dotenv
sqlalchemy~=1.4.46
python-jose
passlib[bcrypt]
python-multipart

alembic~=1.9.2